apps/agent/analyzer/requirements.txt
anthropic
pydantic
python-dotenv
requests
